+++ reductstore/src/storage/query.rs
@@ -463,6 +463,7 @@ mod tests {
             "bucket".to_string(),
             "entry".to_string(),
             Cfg::default().into(),
+            Default::default(),
         )
         .await
         .unwrap();