+++ reductstore/src/storage/query/base.rs
@@ -140,6 +140,7 @@ pub(crate) mod tests {
             "bucket".to_string(),
             "entry".to_string(),
             Cfg::default().into(),
+            Default::default(),
         )
         .await
         .unwrap();